#9 auto setting painless script
