+++ public/app.js
@@ -32,6 +32,7 @@ import MlJobService from './service/mlJob.service';
 import AlertService from './service/alert.service';
 
 import constValue from './const';
+import { script } from './script/script';
 
 uiRoutes.enable();
 uiRoutes
@@ -57,6 +58,7 @@ uiRoutes
 uiModules
   .get('app/ml_alert', [])
   .constant('mlaConst', constValue)
+  .constant('script', script)
   .constant('parse', parse)
   .controller('AlertListController', AlertListController)
   .controller('AlertSettingController', AlertSettingController)
